+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [#1433](https://github.com/oauth2-proxy/oauth2-proxy/pull/1433) Let authentication fail when session validation fails (@stippi2)
 - [#1445](https://github.com/oauth2-proxy/oauth2-proxy/pull/1445) Fix docker container multi arch build issue by passing GOARCH details to make build (@jkandasa)
 - [#1444](https://github.com/oauth2-proxy/oauth2-proxy/pull/1444) Update LinkedIn provider validate URL (@jkandasa)
+- [#1446](https://github.com/oauth2-proxy/oauth2-proxy/pull/1446) Extract roles from keycloak generated tokens when using `--skip-jwt-bearer-tokens=true` (@jplana)
 
 # V7.2.0